feat: deleted some unused ticket components
