--- app/components/shared/ticket-card/accent-decoration.tsx
@@ -1,10 +1,0 @@
-export const AccentDecoration = () => (
-  <div className="absolute -right-10 -bottom-10 w-[280px] h-[300px] pointer-events-none">
-    <img
-      src="/svg/accent.svg"
-      alt=""
-      className="w-full h-full object-contain"
-      aria-hidden="true"
-    />
-  </div>
-);
