--- app/components/shared/ticket-card/ticket-bg-blue.tsx
@@ -1,27 +1,0 @@
-export const TicketBgBlue = () => {
-	return (
-		<svg
-			role="img"
-			aria-label="ticket"
-			width="100%"
-			height="100%"
-			viewBox="0 0 533 388"
-			fill="none"
-			xmlns="http://www.w3.org/2000/svg"
-		>
-			<mask id="path-1-inside-1_902_1928" fill="white">
-				<path d="M513 0C524.046 0 533 8.9543 533 20V46.9297C531.716 46.7262 530.397 46.6191 529.052 46.6191C515.969 46.6194 505.363 56.6385 505.363 68.9971C505.363 81.3555 515.969 91.3738 529.052 91.374C530.397 91.374 531.716 91.266 533 91.0625V296.811C531.716 296.607 530.397 296.5 529.052 296.5C515.969 296.5 505.363 306.519 505.363 318.877C505.363 331.236 515.969 341.255 529.052 341.255C530.397 341.255 531.716 341.147 533 340.943V367.874C533 378.92 524.046 387.874 513 387.874H20C8.95431 387.874 0 378.92 0 367.874V341.255C13.083 341.255 23.6895 331.236 23.6895 318.877C23.6893 306.712 13.4124 296.813 0.612305 296.507L0 296.5V91.374C13.0829 91.374 23.6893 81.3556 23.6895 68.9971C23.6895 56.6384 13.083 46.6191 0 46.6191V20C9.2773e-06 8.95431 8.95431 0 20 0H513Z" />
-			</mask>
-			<path
-				d="M513 0C524.046 0 533 8.9543 533 20V46.9297C531.716 46.7262 530.397 46.6191 529.052 46.6191C515.969 46.6194 505.363 56.6385 505.363 68.9971C505.363 81.3555 515.969 91.3738 529.052 91.374C530.397 91.374 531.716 91.266 533 91.0625V296.811C531.716 296.607 530.397 296.5 529.052 296.5C515.969 296.5 505.363 306.519 505.363 318.877C505.363 331.236 515.969 341.255 529.052 341.255C530.397 341.255 531.716 341.147 533 340.943V367.874C533 378.92 524.046 387.874 513 387.874H20C8.95431 387.874 0 378.92 0 367.874V341.255C13.083 341.255 23.6895 331.236 23.6895 318.877C23.6893 306.712 13.4124 296.813 0.612305 296.507L0 296.5V91.374C13.0829 91.374 23.6893 81.3556 23.6895 68.9971C23.6895 56.6384 13.083 46.6191 0 46.6191V20C9.2773e-06 8.95431 8.95431 0 20 0H513Z"
-				fill="#C4C4C4"
-				fillOpacity="0.6"
-			/>
-			<path
-				d="M533 46.9297L532.217 51.8681L538 52.7844V46.9297H533ZM529.052 46.6191V41.6191H529.052L529.052 46.6191ZM505.363 68.9971H500.363V68.9971L505.363 68.9971ZM529.052 91.374L529.052 96.374H529.052V91.374ZM533 91.0625H538V85.2077L532.217 86.1241L533 91.0625ZM533 296.811L532.217 301.749L538 302.665V296.811H533ZM529.052 296.5V291.5H529.052L529.052 296.5ZM505.363 318.877L500.363 318.877V318.877H505.363ZM529.052 341.255L529.052 346.255H529.052V341.255ZM533 340.943H538V335.089L532.217 336.005L533 340.943ZM0 341.255V336.255H-5V341.255H0ZM23.6895 318.877H28.6895V318.877L23.6895 318.877ZM0.612305 296.507L0.732024 291.508L0.700077 291.508L0.668123 291.507L0.612305 296.507ZM0 296.5H-5V301.444L-0.0558179 301.5L0 296.5ZM0 91.374V86.374H-5V91.374H0ZM23.6895 68.9971L28.6895 68.9971V68.9971H23.6895ZM0 46.6191H-5V51.6191H0V46.6191ZM0 20L-5 20V20H0ZM513 0V5C521.284 5 528 11.7157 528 20H533H538C538 6.19288 526.807 -5 513 -5V0ZM533 20H528V46.9297H533H538V20H533ZM533 46.9297L533.783 41.9913C532.243 41.7474 530.663 41.6191 529.052 41.6191V46.6191V51.6191C530.131 51.6191 531.189 51.7051 532.217 51.8681L533 46.9297ZM529.052 46.6191L529.052 41.6191C513.479 41.6194 500.363 53.6128 500.363 68.9971H505.363H510.363C510.363 59.6642 518.458 51.6193 529.052 51.6191L529.052 46.6191ZM505.363 68.9971L500.363 68.9971C500.363 84.3814 513.48 96.3738 529.052 96.374L529.052 91.374L529.052 86.374C518.458 86.3739 510.363 78.3295 510.363 68.997L505.363 68.9971ZM529.052 91.374V96.374C530.665 96.374 532.246 96.2444 533.783 96.0009L533 91.0625L532.217 86.1241C531.186 86.2875 530.129 86.374 529.052 86.374V91.374ZM533 91.0625H528V296.811H533H538V91.0625H533ZM533 296.811L533.783 291.872C532.243 291.628 530.663 291.5 529.052 291.5V296.5V301.5C530.131 301.5 531.189 301.586 532.217 301.749L533 296.811ZM529.052 296.5L529.052 291.5C513.48 291.5 500.363 303.493 500.363 318.877L505.363 318.877L510.363 318.877C510.363 309.544 518.458 301.5 529.052 301.5L529.052 296.5ZM505.363 318.877H500.363C500.363 334.261 513.479 346.255 529.052 346.255L529.052 341.255L529.052 336.255C518.458 336.255 510.363 328.21 510.363 318.877H505.363ZM529.052 341.255V346.255C530.665 346.255 532.246 346.125 533.783 345.882L533 340.943L532.217 336.005C531.186 336.168 530.129 336.255 529.052 336.255V341.255ZM533 340.943H528V367.874H533H538V340.943H533ZM533 367.874H528C528 376.158 521.284 382.874 513 382.874V387.874V392.874C526.807 392.874 538 381.681 538 367.874H533ZM513 387.874V382.874H20V387.874V392.874H513V387.874ZM20 387.874V382.874C11.7157 382.874 5 376.158 5 367.874H0H-5C-5 381.681 6.19288 392.874 20 392.874V387.874ZM0 367.874H5V341.255H0H-5V367.874H0ZM0 341.255V346.255C15.5723 346.255 28.6895 334.262 28.6895 318.877H23.6895H18.6895C18.6895 328.21 10.5937 336.255 0 336.255V341.255ZM23.6895 318.877L28.6895 318.877C28.6893 303.735 15.9767 291.873 0.732024 291.508L0.612305 296.507L0.492585 301.505C10.8481 301.753 18.6893 309.688 18.6895 318.877L23.6895 318.877ZM0.612305 296.507L0.668123 291.507L0.0558179 291.5L0 296.5L-0.0558179 301.5L0.556487 301.507L0.612305 296.507ZM0 296.5H5V91.374H0H-5V296.5H0ZM0 91.374V96.374C15.572 96.374 28.6892 84.3817 28.6895 68.9971L23.6895 68.9971L18.6895 68.997C18.6893 78.3295 10.5938 86.374 0 86.374V91.374ZM23.6895 68.9971H28.6895C28.6895 53.6125 15.5723 41.6191 0 41.6191V46.6191V51.6191C10.5937 51.6191 18.6895 59.6643 18.6895 68.9971H23.6895ZM0 46.6191H5V20H0H-5V46.6191H0ZM0 20L5 20C5.00001 11.7157 11.7157 5 20 5V0V-5C6.19288 -5 -4.99999 6.19289 -5 20L0 20ZM20 0V5H513V0V-5H20V0Z"
-				fill="#224083"
-				mask="url(#path-1-inside-1_902_1928)"
-			/>
-		</svg>
-	);
-};
