--- app/components/shared/ticket-card/ticket-bg-gray.tsx
@@ -1,27 +1,0 @@
-export const TicketBgGray = () => {
-	return (
-		<svg
-			role="img"
-			aria-label="ticket"
-			width="100%"
-			height="100%"
-			viewBox="0 0 533 388"
-			fill="none"
-			xmlns="http://www.w3.org/2000/svg"
-		>
-			<mask id="path-1-inside-1_901_992" fill="white">
-				<path d="M513 0C524.046 0 533 8.95431 533 20V46.9297C531.716 46.7262 530.397 46.6191 529.052 46.6191C515.969 46.6194 505.363 56.6385 505.363 68.9971C505.363 81.3555 515.969 91.3738 529.052 91.374C530.397 91.374 531.716 91.266 533 91.0625V296.811C531.716 296.607 530.397 296.5 529.052 296.5C515.969 296.5 505.363 306.519 505.363 318.877C505.363 331.236 515.969 341.255 529.052 341.255C530.397 341.255 531.716 341.147 533 340.943V367.874C533 378.92 524.046 387.874 513 387.874H20C8.95431 387.874 0 378.92 0 367.874V341.255C13.083 341.255 23.6895 331.236 23.6895 318.877C23.6893 306.712 13.4124 296.813 0.612305 296.507L0 296.5V91.374C13.0829 91.374 23.6893 81.3556 23.6895 68.9971C23.6895 56.6384 13.083 46.6191 0 46.6191V20C8.97644e-06 8.95431 8.95431 1.00666e-08 20 0H513Z" />
-			</mask>
-			<path
-				d="M513 0C524.046 0 533 8.95431 533 20V46.9297C531.716 46.7262 530.397 46.6191 529.052 46.6191C515.969 46.6194 505.363 56.6385 505.363 68.9971C505.363 81.3555 515.969 91.3738 529.052 91.374C530.397 91.374 531.716 91.266 533 91.0625V296.811C531.716 296.607 530.397 296.5 529.052 296.5C515.969 296.5 505.363 306.519 505.363 318.877C505.363 331.236 515.969 341.255 529.052 341.255C530.397 341.255 531.716 341.147 533 340.943V367.874C533 378.92 524.046 387.874 513 387.874H20C8.95431 387.874 0 378.92 0 367.874V341.255C13.083 341.255 23.6895 331.236 23.6895 318.877C23.6893 306.712 13.4124 296.813 0.612305 296.507L0 296.5V91.374C13.0829 91.374 23.6893 81.3556 23.6895 68.9971C23.6895 56.6384 13.083 46.6191 0 46.6191V20C8.97644e-06 8.95431 8.95431 1.00666e-08 20 0H513Z"
-				fill="white"
-			/>
-			<path
-				d="M533 46.9297L532.687 48.905L535 49.2716V46.9297H533ZM529.052 46.6191V44.6191H529.052L529.052 46.6191ZM505.363 68.9971H503.363V68.9971L505.363 68.9971ZM529.052 91.374L529.052 93.374H529.052V91.374ZM533 91.0625H535V88.7206L532.687 89.0872L533 91.0625ZM533 296.811L532.687 298.786L535 299.152V296.811H533ZM529.052 296.5V294.5H529.052L529.052 296.5ZM505.363 318.877L503.363 318.877V318.877H505.363ZM529.052 341.255L529.052 343.255H529.052V341.255ZM533 340.943H535V338.601L532.687 338.968L533 340.943ZM513 387.874V389.874V387.874ZM0 341.255V339.255H-2V341.255H0ZM23.6895 318.877H25.6895V318.877L23.6895 318.877ZM0.612305 296.507L0.660192 294.507L0.647414 294.507L0.634632 294.507L0.612305 296.507ZM0 296.5H-2V298.478L-0.0223272 298.5L0 296.5ZM0 91.374V89.374H-2V91.374H0ZM23.6895 68.9971L25.6895 68.9971V68.9971H23.6895ZM0 46.6191H-2V48.6191H0V46.6191ZM0 20L-2 20V20H0ZM20 0V-2V0ZM513 0V2C522.941 2 531 10.0589 531 20H533H535C535 7.84974 525.15 -2 513 -2V0ZM533 20H531V46.9297H533H535V20H533ZM533 46.9297L533.313 44.9543C531.927 44.7347 530.503 44.6191 529.052 44.6191V46.6191V48.6191C530.291 48.6191 531.505 48.7178 532.687 48.905L533 46.9297ZM529.052 46.6191L529.052 44.6191C514.973 44.6194 503.363 55.4282 503.363 68.9971H505.363H507.363C507.363 57.8488 516.965 48.6193 529.052 48.6191L529.052 46.6191ZM505.363 68.9971L503.363 68.9971C503.363 82.5659 514.973 93.3738 529.052 93.374L529.052 91.374L529.052 89.374C516.965 89.3738 507.363 80.1451 507.363 68.997L505.363 68.9971ZM529.052 91.374V93.374C530.504 93.374 531.928 93.2574 533.313 93.0378L533 91.0625L532.687 89.0872C531.504 89.2746 530.29 89.374 529.052 89.374V91.374ZM533 91.0625H531V296.811H533H535V91.0625H533ZM533 296.811L533.313 294.835C531.927 294.616 530.503 294.5 529.052 294.5V296.5V298.5C530.291 298.5 531.505 298.599 532.687 298.786L533 296.811ZM529.052 296.5L529.052 294.5C514.973 294.5 503.363 305.308 503.363 318.877L505.363 318.877L507.363 318.877C507.363 307.729 516.965 298.5 529.052 298.5L529.052 296.5ZM505.363 318.877H503.363C503.363 332.446 514.973 343.255 529.052 343.255L529.052 341.255L529.052 339.255C516.965 339.255 507.363 330.025 507.363 318.877H505.363ZM529.052 341.255V343.255C530.504 343.255 531.928 343.138 533.313 342.919L533 340.943L532.687 338.968C531.504 339.155 530.29 339.255 529.052 339.255V341.255ZM533 340.943H531V367.874H533H535V340.943H533ZM533 367.874H531C531 377.815 522.941 385.874 513 385.874V387.874V389.874C525.15 389.874 535 380.024 535 367.874H533ZM513 387.874V385.874H20V387.874V389.874H513V387.874ZM20 387.874V385.874C10.0589 385.874 2 377.815 2 367.874H0H-2C-2 380.024 7.84974 389.874 20 389.874V387.874ZM0 367.874H2V341.255H0H-2V367.874H0ZM0 341.255V343.255C14.0787 343.255 25.6895 332.446 25.6895 318.877H23.6895H21.6895C21.6895 330.025 12.0873 339.255 0 339.255V341.255ZM23.6895 318.877L25.6895 318.877C25.6893 305.521 14.4381 294.837 0.660192 294.507L0.612305 296.507L0.564417 298.506C12.3867 298.789 21.6893 307.902 21.6895 318.877L23.6895 318.877ZM0.612305 296.507L0.634632 294.507L0.0223272 294.5L0 296.5L-0.0223272 298.5L0.589978 298.507L0.612305 296.507ZM0 296.5H2V91.374H0H-2V296.5H0ZM0 91.374V93.374C14.0785 93.374 25.6893 82.5661 25.6895 68.9971L23.6895 68.9971L21.6895 68.997C21.6893 80.1451 12.0873 89.374 0 89.374V91.374ZM23.6895 68.9971H25.6895C25.6895 55.428 14.0787 44.6191 0 44.6191V46.6191V48.6191C12.0873 48.6191 21.6895 57.8488 21.6895 68.9971H23.6895ZM0 46.6191H2V20H0H-2V46.6191H0ZM0 20L2 20C2.00001 10.0589 10.0589 2 20 2V0V-2C7.84974 -2 -1.99999 7.84974 -2 20L0 20ZM20 0V2H513V0V-2H20V0Z"
-				fill="#333333"
-				fillOpacity="0.2"
-				mask="url(#path-1-inside-1_901_992)"
-			/>
-		</svg>
-	);
-};
